requirement.txt
fastapi
uvicorn[standard]
sqlmodel
requests
google-cloud-firestore
py_eureka_client
marshmallow-sqlalchemy==0.30.0
python-dotenv==1.0.0
pymysql==1.1.0